Add stack implementation using linked list
Added stack implementation using a linked list.
This can be achieved in java by creating a inner class (private class)
that containts a filed that holds the data and another one that points
to the next.
Push and pop operation are taking a constant time of one, but
the downside is that more space is required for holding the pointer
to the next element